fix(tests): fold the identity tests into the module's own mirror (PS-204 orphan-test-file)
CI caught my second audit violation of the evening, and it was right again.

I moved the test to tests/scitex_agent_container/runtimes/ to satisfy PS-302
(tests-unknown-subdir), then tripped PS-204 (orphan-test-file): a test file named
test_identity_is_agent_scoped_not_directory_scoped.py mirrors NO source module.
The mirror is per-MODULE — src/.../runtimes/_cct_token_pool.py maps to
tests/.../runtimes/test__cct_token_pool.py, and nothing else.

I had read the rule as "descriptive names" and applied it to the FILE. It belongs
on the FUNCTIONS: test_two_agents_in_one_repo_get_different_identities says what
broke; the file says which module owns it. Both rules are satisfied at once, and
they were never in tension — I invented the tension by picking the wrong noun.

So the 8 identity tests now live in test__cct_token_pool.py under a section
header carrying the incident, and the orphan file is gone. No test lost, no
assertion weakened, names unchanged.

Also caught by running them rather than trusting the move: the merge dropped
_default_agent_id from the import list — 2 NameErrors. Fixed and re-run.

31/31 pass.
